refactor: Move `sha256OrNull` to PM and fix loop cancellation
